requirements.txt
pywebview
Flask
requests
yt-dlp
